// src/app/core/models/model/userProfileResponse.ts
/**
 * Wrapity
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UserProfileResponse { 
    id: number;
    username: string;
    bio?: string | null;
    avatar_url?: string | null;
    reviews_count: number;
    followers_count: number;
    following_count: number;
    is_following?: boolean | null;
}